+++ application/views/Help.php
@@ -107,5 +107,6 @@
 		<li>Yummy Gummy Scans</li>
 		<li>Champion Scans</li>
 		<li>Pure Mashiro Scans</li>
+		<li>Ravens Scans</li>
 	</ul>
 </div>